Restructuring terminal interface
